finished refactoring and styling
