+++ src/app/app-providers.jsx
@@ -1,10 +1,8 @@
 import { ThemeProvider } from 'features/theme';
 import { TodosProvider } from 'features/todos';
 
-export const AppProviders = ({ children }) => {
-  return (
-    <ThemeProvider>
-      <TodosProvider>{children}</TodosProvider>
-    </ThemeProvider>
-  );
-};
+export const AppProviders = ({ children }) => (
+  <ThemeProvider>
+    <TodosProvider>{children}</TodosProvider>
+  </ThemeProvider>
+);
