+++ src/features/layout/page/theme-switch.module.css
@@ -1,3 +1,3 @@
 .action img {
-  height: 0.6em;
+  height: 0.65em;
 }
