--- src/features/todos/todo-container/todo-container.module.css
@@ -1,6 +1,0 @@
-.footerMessage {
-  margin-top: 2em;
-  text-align: center;
-  color: var(--status-text-color);
-  font-size: 0.75rem;
-}
