--- src/features/todos/todo-container/todo-filter.module.css
@@ -1,34 +1,0 @@
-@value bp-mobile-screen from 'shared/styles/breakpoints.module.css';
-
-.filter {
-  grid-area: filter;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1em;
-  justify-content: center;
-  align-content: center;
-  padding-inline: 1em;
-  font-weight: var(--font-weight-bold);
-}
-@media bp-mobile-screen {
-  .filter {
-    gap: 2em;
-    margin-top: 1.3rem;
-    border-radius: 5px;
-  }
-}
-
-.filter input[type='radio'] {
-  display: none;
-}
-.filter input[type='radio'] + label {
-  cursor: pointer;
-  font-size: 0.9375rem;
-}
-.filter input[type='radio']:hover + label,
-.filter input[type='radio']:focus + label {
-  color: var(--todo-text-color);
-}
-.filter input[type='radio']:checked + label {
-  color: var(--color-bright-blue);
-}
