+++ src/features/todos/todo-dragndrop-directives/todo-dragndrop-directives.jsx
@@ -1,0 +1,5 @@
+import styles from './todo-dragndrop-directives.module.css';
+
+export const DragNDropDirectives = () => (
+  <p className={styles.message}>(not yet) Drag and drop to reorder list</p>
+);
