+++ src/features/todos/todo-entry/index.js
@@ -1,0 +1,1 @@
+export { TodoEntry } from './todo-entry';
