+++ src/features/todos/todo-entry/todo-entry.module.css
@@ -1,6 +1,5 @@
 .entry {
   background: var(--todo-background-color);
-  color: var(--todo-text-color);
   border-radius: 5px;
-  margin-block: 1em;
+  margin-block: 1.2em;
 }
