+++ src/features/todos/todo-item/checkbox.module.css
@@ -1,28 +1,34 @@
 :root {
-  --todo-checkbox-width: 1em;
+  --todo-checkbox-width: 1.25em;
 }
+
 input[type='checkbox'].checkbox {
   display: none;
 }
+
 input[type='checkbox'].checkbox + label {
   position: relative;
-  padding-left: calc(var(--todo-checkbox-width) + 1em);
+  min-height: var(--todo-checkbox-width); /*to deal with empty labels*/
+  padding-left: calc(var(--todo-checkbox-width) + 1.3em);
 }
 input[type='checkbox'].checkbox:not(:disabled) + label {
   cursor: pointer;
 }
+
 input[type='checkbox'].checkbox:checked + label {
   text-decoration: line-through;
   color: var(--todo-completed-text-color);
 }
+
 input[type='checkbox'].checkbox:hover:not(:disabled):not(:checked) + label,
 input[type='checkbox'].checkbox:focus:not(:disabled):not(:checked) + label {
   color: var(--todo-hover-text-color);
 }
+
 input[type='checkbox'].checkbox + label::before,
 input[type='checkbox'].checkbox + label::after {
   position: absolute;
-  top: 0;
+  top: calc((1em - var(--todo-checkbox-width)) / 2);
   left: 0;
   display: inline-block;
   content: '';
@@ -44,7 +50,7 @@ input[type='checkbox'].checkbox:checked + label::after {
 }
 input[type='checkbox'].checkbox:hover:not(:disabled):not(:checked) + label::after,
 input[type='checkbox'].checkbox:focus:not(:disabled):not(:checked) + label::after {
-  top: 1px;
+  top: calc(((1em - var(--todo-checkbox-width)) / 2) + 1px);
   left: 1px;
   width: calc(var(--todo-checkbox-width) - 2px);
   background: var(--todo-background-color);
