+++ src/features/todos/todo-item/text-input.module.css
@@ -2,7 +2,7 @@
   width: 100%;
   height: 100%;
   border: none;
-  background: inherit;
+  background: transparent;
   font: inherit;
   color: inherit;
 }
