+++ src/features/todos/todo-list/index.js
@@ -1,0 +1,1 @@
+export { TodoList } from './todo-list';
