+++ src/shared/ui/actionnable.jsx
@@ -2,11 +2,7 @@ import styles from './actionnable.module.css';
 
 export const Actionnable = ({ ...props }) => {
   return (
-    <button
-      {...props}
-      className={`${styles.button} ${props.className ?? ''}`}
-      onClick={props.onClick}
-    >
+    <button {...props} className={`${styles.button} ${props.className ?? ''}`}>
       {props.children}
     </button>
   );
