+++ src/shared/utils/helpers.js
@@ -1,3 +1,2 @@
-export const getRandomHtmlId = (prefix = 'i') => {
-  return `${prefix}${Math.trunc(Math.random() * 10 * 1000000 * 1000000000)}`;
-};
+export const getRandomHtmlId = (prefix = 'i') =>
+  `${prefix}${Math.trunc(Math.random() * 10 * 1000000 * 1000000000)}`;
